pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'viet-asr'
version = '0.1.0'
description = 'Vietnamese speech-to-text pipeline with pretrained model fine-tuning and streaming inference.'
authors = [ { name = 'NhapMonAI Team' } ]
readme = 'README.md'
requires-python = '>=3.10'
dependencies = [
    'torch>=2.1',
    'torchaudio>=2.1',
    'transformers>=4.36',
    'datasets>=2.18',
    'librosa>=0.10',
    'soundfile>=0.12',
    'accelerate>=0.27',
    'jiwer>=3.0',
    'pydantic>=2.7',
    'pyyaml>=6.0',
    'typer>=0.12',
    'fastapi>=0.110',
    'uvicorn[standard]>=0.27',
    'websockets>=12.0',
    'pyttsx3>=2.90'
]

[tool.setuptools]
packages = ['asr']

